move set_nonblock, create_socket and path_exists to libi3
